requirements.txt
streamlit
requests
python-dotenv
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
musicbrainzngs
discogs-client
